% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_vldt_trggr_tbl_vldtn_10.R
\name{ct_vldt_trggr_tbl_vldtn_10}
\alias{ct_vldt_trggr_tbl_vldtn_10}
\title{<Add Title>}
\usage{
ct_vldt_trggr_tbl_vldtn_10(filePath, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
<Add Description>
}
