% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantheraIDSImageViewer.R
\name{imageViewer}
\alias{imageViewer}
\title{Image Viewer Function}
\usage{
imageViewer(filePath, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{filePath}{A string path to a csv file which contains the images source's}

\item{width}{Fixed width for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{elementId}{Use an explicit element ID for the widget
 Useful if you have other JavaScript that needs to
 explicitly discover and interact with a specific widget instance.
in any other case leave as NULL which results in an
automatically generated one.}
}
\description{
This function maps images to a html component to complete the
             rendering of images...
}
\details{
Click Events And Keyboard Shortcuts
           \itemize{
           \item{click - }{Views an image}
           \item{shiftKey + click - }{selects / mark an image.}
           \item{left arrow - }{View the previous image.}
           \item{right arrow - }{View the next image.}
           \item{up arrow - }{Zoom in the image.}
           \item{down arrow - }{Zoom out the image.}
           \item{Ctrl + 0 - }{Zoom out to initial size.}
           \item{Ctrl + 1 - }{Zoom in to natural size.}
           \item{Esc - }{ Exit full screen or stop play.}
             }
}
\examples{
imageViewer('./www/data/myfile.csv')

}
\keyword{viewer.js}
