% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantheraIDSImageViewer-package.R
\docType{package}
\name{PantheraIDSImageViewer-package}
\alias{PantheraIDSImageViewer}
\alias{PantheraIDSImageViewer-package}
\title{PantheraIDSImageViewer: Image Viewer Widget}
\description{
Views images in a title pattern .
}
\author{
\strong{Maintainer}: Valentine Tawira \email{vtawira@panthera.org} (YOUR-ORCID-ID)

}
\keyword{internal}
