% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantheraIDSImageViewer.R
\name{PantheraIDSImageViewer}
\alias{PantheraIDSImageViewer}
\title{PantheraIDSImageViewer}
\usage{
PantheraIDSImageViewer(filePath, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
Takes a path to a file and renders an image viewer
}
