% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pttrn_rcgntn_img_vwr.R
\name{pttrn_rcgntn_img_vwr-shiny}
\alias{pttrn_rcgntn_img_vwr-shiny}
\alias{pttrn_rcgntn_img_vwrOutput}
\alias{renderPttrn_rcgntn_img_vwr}
\title{Shiny bindings for pttrn_rcgntn_img_vwr}
\usage{
pttrn_rcgntn_img_vwrOutput(outputId, width = "100\%", height = "400px")

renderPttrn_rcgntn_img_vwr(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a pttrn_rcgntn_img_vwr}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using pttrn_rcgntn_img_vwr within Shiny
applications and interactive Rmd documents.
}
