% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pttrn_rcgntn_img_vwr.R
\name{pttrn_rcgntn_img_vwr}
\alias{pttrn_rcgntn_img_vwr}
\title{<Add Title>}
\usage{
pttrn_rcgntn_img_vwr(filePath, width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}
