% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantheraIDSImageViewer.R
\name{imageViewer}
\alias{imageViewer}
\title{Image Viewer Function}
\usage{
imageViewer(filePath, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{filePath}{A path to a csv file which contains the images source path's}

\item{width}{Fixed width for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{elementId}{Use an explicit element ID for the widget
 Useful if you have other JavaScript that needs to
 explicitly discover and interact with a specific widget instance.
in any other case leave as NULL which results in an
automatically generated one.}
}
\description{
This function allows renders images to a html component.
}
\examples{
imageViewer('./www/data/myfile.csv')

}
\keyword{viewer.js}
