% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PantheraIDSImageViewer-package.R
\docType{package}
\name{PantheraIDSImageViewer-package}
\alias{PantheraIDSImageViewer}
\alias{PantheraIDSImageViewer-package}
\title{PantheraIDSImageViewer: View Image Panels with the JavaScript 'viewer.js' Library}
\description{
Image viewer widget for displaying images as tile/thumbnail
}
\author{
\strong{Maintainer}: Valentine Tawira \email{vtawira@panthera.org}

}
\keyword{internal}
